Move helper classes out of concurrent_hash_map.hpp file
