+++ include/libpmemobj++/detail/common.hpp
@@ -43,6 +43,11 @@
 #include <string>
 #include <typeinfo>
 
+#if _MSC_VER
+#include <intrin.h>
+#include <windows.h>
+#endif
+
 #if defined(__GNUC__) || defined(__clang__)
 #define POBJ_CPP_DEPRECATED __attribute__((deprecated))
 #elif defined(_MSC_VER)
@@ -201,6 +206,42 @@ next_pow_2(uint32_t v)
 	++v;
 	return v + (v == 0);
 }
+
+#if _MSC_VER
+static inline int
+Log2(uint64_t x)
+{
+	unsigned long j;
+	_BitScanReverse64(&j, x);
+	return static_cast<int>(j);
+}
+#elif __GNUC__ || __clang__
+static inline int
+Log2(uint64_t x)
+{
+	// __builtin_clz builtin count _number_ of leading zeroes
+	return 8 * int(sizeof(x)) - __builtin_clzll(x) - 1;
+}
+#else
+static inline int
+Log2(uint64_t x)
+{
+	x |= (x >> 1);
+	x |= (x >> 2);
+	x |= (x >> 4);
+	x |= (x >> 8);
+	x |= (x >> 16);
+	x |= (x >> 32);
+
+	static const int table[64] = {
+		0,  58, 1,  59, 47, 53, 2,  60, 39, 48, 27, 54, 33, 42, 3,  61,
+		51, 37, 40, 49, 18, 28, 20, 55, 30, 34, 11, 43, 14, 22, 4,  62,
+		57, 46, 52, 38, 26, 32, 41, 50, 36, 17, 19, 29, 10, 13, 21, 56,
+		45, 25, 31, 35, 16, 9,  12, 44, 24, 15, 8,  23, 7,  6,  5,  63};
+
+	return table[(x * 0x03f6eaf2cd271461) >> 58];
+}
+#endif
 
 } /* namespace detail */
